fix(tests): split Pest.php helpers into Support/helpers.php
Root-cause of the persistent "Call to undefined function
Cassandra\Tests\Feature\Tuples\migrateKeyspace()" error:

When phpunit.xml's bootstrap requires vendor/autoload.php, composer
must somehow load Pest.php. Pest 3's auto-discovery walks up from
each test file looking for Pest.php in parent directories, but is
brittle to the working-directory and the presence of the nested
tests/composer.json.

Putting Pest.php at repo root or adding it to composer's
autoload-dev.files both fail because Pest.php calls uses() and
expect()->extend() at file scope — and those Pest functions are not
yet defined when composer's classloader runs the autoload-dev.files
list. Result: fatal "Call to undefined function uses()".

Fix: separate concerns.
- tests/Support/helpers.php  — env, scyllaDbConnection, migrateKeyspace,
  dropKeyspace. No Pest API used. Loaded eagerly via composer's
  autoload-dev.files; available in any test file regardless of
  Pest's bootstrap state.
- tests/Pest.php             — only the Pest-specific calls (uses() group
  registration, expect()->extend('map')). Pest auto-discovers it from
  the conventional tests/ location.

Confirmed locally: namespaced callers like
  namespace Cassandra\Tests\Feature\Tuples;
  migrateKeyspace(...)
fall through to the global helper as intended.

Reverts the previous attempt to relocate Pest.php / phpunit.xml to
the repo root — that move was the wrong layer to fix the problem.
Both files return to their canonical Pest 3 location at tests/.

The env() helper now consults \$_SERVER and \$_ENV before getenv() so
phpunit.xml <php><env/></php> blocks (and Symfony-style overrides)
take effect.
