+++ scripts/run-tests.sh
@@ -16,5 +16,6 @@ ninja install || exit 1
 
 popd || exit 1
 cp cassandra.ini /usr/local/etc/php/conf.d/cassandra.ini
-( cd tests && composer install )
-php ./tests/vendor/bin/pest
+cd tests || exit 1
+composer install
+php ./vendor/bin/pest
